small bug

+++ src/MessageCore/Serializing/JsonMessageSerializer.cs
@@ -116,7 +116,7 @@ namespace VVVV.Packs.Messaging.Serializing
                     switch (jType)
                     {
                         case JTokenType.Boolean: itemType = typeof(bool); break;
-                        case JTokenType.Float: itemType = typeof(float); break;
+                        case JTokenType.Float: itemType = typeof(double); break;
                         case JTokenType.Integer: itemType = typeof(int); break;
                         case JTokenType.String: itemType = typeof(string); break;
